use static agent icons in Zellij tabs
Zellij status animation required a background process per working tab and
continuously spawned CLI commands to read and rename tabs. This imposed
substantial process, IPC, redraw, and lifecycle overhead for a cosmetic effect.

Render the configured working, waiting, and done icons when status hooks fire,
matching tmux behavior. Keep stable tab-ID targeting, marker-based suffix
replacement, tmux style stripping, and canonical tab-name handling. Remove the
spinner command, runtime state, socket validation, and animation process.

Zellij waiting and done icons remain visible until the next status update
because Zellij does not provide an equivalent to tmux's focus hook.
