+++ src/cli.rs
@@ -718,19 +718,6 @@ enum Commands {
         command: command::set_window_status::SetWindowStatusCommand,
     },
 
-    /// Animate a Zellij tab's working status (internal use)
-    #[command(hide = true, name = "_zellij-status-spinner")]
-    ZellijStatusSpinner {
-        #[arg(long)]
-        session: String,
-
-        #[arg(long)]
-        tab_id: u32,
-
-        #[arg(long)]
-        token: String,
-    },
-
     /// Set the base branch for the current worktree (used after rebasing)
     #[command(hide = true, name = "set-base")]
     SetBase {
@@ -1110,11 +1097,6 @@ pub fn run() -> Result<()> {
         },
         Commands::Sandbox(args) => command::sandbox::run(args),
         Commands::SetWindowStatus { command } => command::set_window_status::run(command),
-        Commands::ZellijStatusSpinner {
-            session,
-            tab_id,
-            token,
-        } => crate::multiplexer::zellij_status::run_spinner(&session, tab_id, &token),
         Commands::SetBase { base } => command::set_base::run(&base),
         Commands::LastDone => command::last_done::run(),
         Commands::LastAgent => command::last_agent::run(),
